Add files via upload
Dashboard Owner: Enakhe Ewanlen
Data sources: Makeover Monday 

This dashboard shows the loan amount disbursed between different sectors (Corporate, Individual, and Partnership business) among different states in the United States.
It can be seen that the corporate sector has the highest loan approval, followed by the individual sector, and lastly, the partnership business sector.  Wells Fargo bank is the top borrowing bank with 14,039,261,400 total loans across all three sectors. However, it can also be seen that it is not the top borrowing bank in the individual and partnership sectors of the economy, which Live Oak Banking holds.
Also, across the states, CA has the highest loan approval as it can be seen it has 16.40% of all the loans disbursed across the country.